Updated structure of project and created separate router for projects endpoints
