+++ utils/logger.ts
@@ -1,0 +1,20 @@
+import winston from "winston";
+
+const logger = winston.createLogger({
+    // Log only if level is less than (meaning more severe) or equal to this
+    level: "info",
+    // Use timestamp and printf to create a standard log format
+    format: winston.format.combine(
+        winston.format.timestamp(),
+        winston.format.printf(
+            (info) => `${info.timestamp} ${info.level}: ${info.message}`
+        )
+    ),
+    // Log to the console and a file
+    transports: [
+        new winston.transports.Console(),
+        new winston.transports.File({ filename: "logs/app.log" }),
+    ],
+});
+
+export default logger;
